Remove ThesisTable component and create two new components
ThesisTable component was removed and replaced with two new components - ActiveThesisTable and ApplicantsTable.
This was done to split the functionality and make the code more modular.
ActiveThesisTable component is responsible for displaying the list of active theses, while ApplicantsTable manages the list of applicants for the theses.
Testing and styles were also updated according to the changes.